+++ src/compounds/legacy-product-table/src/stories.tsx
@@ -85,7 +85,7 @@ const ExampleProductTable = () => {
         <img src="https://placekitten.com/42/75?image=9" alt="Salman" />
       </ImageCell>
 
-      <CtaCell href={clickableRow}>See Deal</CtaCell>
+      <CtaCell href={''}>See Deal</CtaCell>
     </LegacyProductTable>
   )
 }
@@ -110,10 +110,73 @@ Example.story = {
   }
 }
 
+const DisabledProductTable = () => {
+  return (
+    <LegacyProductTable
+      representativeExample={repExample}
+      repExampleLabel="Representative example: "
+      info={info}
+      title={title}
+      eligibilityContent={eligibilityContent}
+      clickableRow={clickableRow}
+      onClickEligibility={onClickEligibility}
+      eligibilityAddon={eligibilityAddon}
+      disabled
+    >
+      <DataCell
+        backgroundColor="#fef6ed"
+        borderBottomColor="#fcdbb7"
+        label="Loan amount"
+      >
+        £1,000 to £10,000
+      </DataCell>
+
+      <DataCell
+        backgroundColor="#fef1ec"
+        borderBottomColor="#fac9b0"
+        label="Representative APR"
+      >
+        29% APR (£1,000 to £10,000)
+      </DataCell>
+
+      <DataCell
+        backgroundColor="#fbedf1"
+        borderBottomColor="#f0b9c8"
+        label="Loan term"
+      >
+        1 year to 5 years
+      </DataCell>
+
+      <ImageCell>
+        <img src="https://placekitten.com/42/75?image=9" alt="Salman" />
+      </ImageCell>
+
+      <CtaCell href={clickableRow} disabled>
+        See Deal
+      </CtaCell>
+    </LegacyProductTable>
+  )
+}
+
+export const DisabledExample = () => {
+  return (
+    <div>
+      <DisabledProductTable />
+    </div>
+  )
+}
+
+DisabledExample.story = {
+  parameters: {
+    percy: { skip: true }
+  }
+}
+
 export const AutomatedTests = () => {
   return (
     <AllThemes>
       <Example />
+      <DisabledExample />
     </AllThemes>
   )
 }