+++ src/compounds/hero-card/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.15](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-card@2.0.13...@uswitch/trustyle.hero-card@2.0.15) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.hero-card
+
+
+
+
+
+## [2.0.14](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-card@2.0.13...@uswitch/trustyle.hero-card@2.0.14) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.hero-card
+
+
+
+
+
+## [2.0.13](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-card@2.0.12...@uswitch/trustyle.hero-card@2.0.13) (2020-12-10)
+
+**Note:** Version bump only for package @uswitch/trustyle.hero-card
+
+
+
+
+
 ## [2.0.12](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-card@2.0.11...@uswitch/trustyle.hero-card@2.0.12) (2020-12-10)
 
 **Note:** Version bump only for package @uswitch/trustyle.hero-card