+++ src/compounds/sponsored-product-rate-table/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.15](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.14...@uswitch/trustyle.sponsored-product-rate-table@3.3.15) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.14](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.12...@uswitch/trustyle.sponsored-product-rate-table@3.3.14) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.13](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.12...@uswitch/trustyle.sponsored-product-rate-table@3.3.13) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.12](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.11...@uswitch/trustyle.sponsored-product-rate-table@3.3.12) (2020-12-21)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.11](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.10...@uswitch/trustyle.sponsored-product-rate-table@3.3.11) (2020-12-18)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.10](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.8...@uswitch/trustyle.sponsored-product-rate-table@3.3.10) (2020-12-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.9](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.8...@uswitch/trustyle.sponsored-product-rate-table@3.3.9) (2020-12-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.8](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.7...@uswitch/trustyle.sponsored-product-rate-table@3.3.8) (2020-12-16)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.7](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.6...@uswitch/trustyle.sponsored-product-rate-table@3.3.7) (2020-12-15)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
+## [3.3.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.5...@uswitch/trustyle.sponsored-product-rate-table@3.3.6) (2020-12-10)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table
+
+
+
+
+
 ## [3.3.5](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product-rate-table@3.3.4...@uswitch/trustyle.sponsored-product-rate-table@3.3.5) (2020-12-10)
 
 **Note:** Version bump only for package @uswitch/trustyle.sponsored-product-rate-table