+++ src/elements/breadcrumbs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.breadcrumbs@2.1.5...@uswitch/trustyle.breadcrumbs@2.1.6) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.breadcrumbs
+
+
+
+
+
+## [2.1.5](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.breadcrumbs@2.1.4...@uswitch/trustyle.breadcrumbs@2.1.5) (2020-12-18)
+
+**Note:** Version bump only for package @uswitch/trustyle.breadcrumbs
+
+
+
+
+
+## [2.1.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.breadcrumbs@2.1.3...@uswitch/trustyle.breadcrumbs@2.1.4) (2020-12-15)
+
+**Note:** Version bump only for package @uswitch/trustyle.breadcrumbs
+
+
+
+
+
 ## [2.1.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.breadcrumbs@2.1.2...@uswitch/trustyle.breadcrumbs@2.1.3) (2020-12-02)
 
 **Note:** Version bump only for package @uswitch/trustyle.breadcrumbs