+++ src/elements/button/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.button@3.0.2...@uswitch/trustyle.button@3.0.4) (2021-01-05)
+
+
+### Bug Fixes
+
+* type difference between button elements ([9cb1800](https://github.com/uswitch/trustyle/commit/9cb1800))
+
+
+
+
+
+## [3.0.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.button@3.0.2...@uswitch/trustyle.button@3.0.3) (2021-01-05)
+
+
+### Bug Fixes
+
+* type difference between button elements ([26b5374](https://github.com/uswitch/trustyle/commit/26b5374))
+
+
+
+
+
+## [3.0.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.button@3.0.1...@uswitch/trustyle.button@3.0.2) (2020-12-21)
+
+
+### Bug Fixes
+
+* ts check ([d20edf8](https://github.com/uswitch/trustyle/commit/d20edf8))
+
+
+
+
+
 ## [3.0.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.button@3.0.0...@uswitch/trustyle.button@3.0.1) (2020-11-23)