+++ src/elements/card/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.3...@uswitch/trustyle.card@1.12.4) (2021-01-20)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
+## [1.12.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.2...@uswitch/trustyle.card@1.12.3) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
+## [1.12.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.1...@uswitch/trustyle.card@1.12.2) (2020-12-18)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
+## [1.12.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.0...@uswitch/trustyle.card@1.12.1) (2020-12-15)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
+# [1.12.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.11.3...@uswitch/trustyle.card@1.12.0) (2020-12-15)
+
+
+### Features
+
+* add altered featured variant to card component ([fc7d30c](https://github.com/uswitch/trustyle/commit/fc7d30c))
+
+
+
+
+
 ## [1.11.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.11.2...@uswitch/trustyle.card@1.11.3) (2020-12-02)
 
 **Note:** Version bump only for package @uswitch/trustyle.card