+++ src/elements/modal/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.modal@2.4.5...@uswitch/trustyle.modal@2.4.6) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.modal
+
+
+
+
+
+## [2.4.5](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.modal@2.4.4...@uswitch/trustyle.modal@2.4.5) (2020-12-18)
+
+**Note:** Version bump only for package @uswitch/trustyle.modal
+
+
+
+
+
+## [2.4.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.modal@2.4.3...@uswitch/trustyle.modal@2.4.4) (2020-12-15)
+
+**Note:** Version bump only for package @uswitch/trustyle.modal
+
+
+
+
+
 ## [2.4.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.modal@2.4.2...@uswitch/trustyle.modal@2.4.3) (2020-12-02)
 
 **Note:** Version bump only for package @uswitch/trustyle.modal