+++ src/elements/pagination/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.pagination@1.6.5...@uswitch/trustyle.pagination@1.6.6) (2021-01-05)
+
+**Note:** Version bump only for package @uswitch/trustyle.pagination
+
+
+
+
+
+## [1.6.5](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.pagination@1.6.4...@uswitch/trustyle.pagination@1.6.5) (2020-12-18)
+
+**Note:** Version bump only for package @uswitch/trustyle.pagination
+
+
+
+
+
+## [1.6.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.pagination@1.6.3...@uswitch/trustyle.pagination@1.6.4) (2020-12-15)
+
+**Note:** Version bump only for package @uswitch/trustyle.pagination
+
+
+
+
+
 ## [1.6.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.pagination@1.6.2...@uswitch/trustyle.pagination@1.6.3) (2020-12-02)
 
 **Note:** Version bump only for package @uswitch/trustyle.pagination