+++ src/elements/primary-info-block/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.primary-info-block@1.0.7...@uswitch/trustyle.primary-info-block@1.0.9) (2020-12-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.primary-info-block
+
+
+
+
+
+## [1.0.8](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.primary-info-block@1.0.7...@uswitch/trustyle.primary-info-block@1.0.8) (2020-12-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.primary-info-block
+
+
+
+
+
+## [1.0.7](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.primary-info-block@1.0.6...@uswitch/trustyle.primary-info-block@1.0.7) (2020-12-16)
+
+**Note:** Version bump only for package @uswitch/trustyle.primary-info-block
+
+
+
+
+
 ## [1.0.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.primary-info-block@1.0.5...@uswitch/trustyle.primary-info-block@1.0.6) (2020-11-11)